+++ packages/colony-js-client/src/ColonyNetworkClient/index.js
@@ -182,9 +182,7 @@ export default class ColonyNetworkClient extends ContractClient {
   */
   createToken: ColonyNetworkClient.Sender<
     {
-      name: string, // The name of the token.
       symbol: string, // The symbol of the token.
-      decimals: number, // The number of decimals.
     },
     {},
     ColonyNetworkClient,